requirements.txt
python-dotenv
yt-dlp
faster-whisper
requests
beautifulsoup4
openai
anthropic
google-generativeai
yfinance
